backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.26.0
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
email-validator==2.1.0
fastapi-mail
twilio
aiohttp==3.9.1
pytz==2023.3
telethon==1.35.0
weaviate-client==4.4.0
sentence-transformers==2.2.2
torch==2.1.2
transformers==4.35.0
pypdf==3.17.0
python-docx==1.1.0
cryptography>=41.0.0